requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.1
stripe==8.4.0
pyarmor==8.3.5
cryptography==42.0.4
hyqcopt==0.1.0
Meta-Learn==0.1.0
automl-self-improvement==0.1.0
numpy==1.26.4
scikit-learn==1.4.0
pandas==2.2.1
cachetools==5.3.2
python-jose[cryptography]==3.3.0
# requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.28
sqlalchemy>=2.0
snowflake-connector-python>=3.0
google-cloud-bigquery>=3.0
boto3>=1.26
kaggle>=1.5
dask>=2023.0
pyarrow>=11.0
datasets>=2.0  # Hugging Face
python-dotenv>=0.21